requirements.txt
crewai>=0.11.0
tavily-python>=0.2.6
python-dotenv>=1.0.0
langchain>=0.1.0
openai>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.2
cors>=1.0.1 